# client/public/openapi.yaml
openapi: 3.0.3
info:
  title: Kallia API
  version: "1.1.0"
  description: |
    API HTTP do **Kallia** — plataforma PABX VoIP profissional para WhatsApp com agentes de IA (Gemini Live) e suporte a SaaS Multi-tenant.

    ## Autenticação
    A API suporta múltiplos métodos de autenticação:
    1. **JWT Bearer Token**: Envie no cabeçalho `Authorization: Bearer <TOKEN>` (obtido via `POST /api/auth/login`).
    2. **Chave Mestre de API**: Envie no cabeçalho `X-API-Key` (configurada via `KALLIA_API_KEY`).
    3. **Chave de API da Conexão**: Envie no cabeçalho `X-Connection-API-Key` ou `X-API-Key` (chave individual `kc_...` gerada por WhatsApp).
    4. **Ticket de Uso Único (SSE / WS)**: Obtenha em `POST /api/events/ticket` e passe em `?ticket=` para conexões sem headers.

    > ⚠️ A API inicia e gerencia as chamadas (sinalização VoIP e IA no servidor), enquanto o **áudio manual via navegador** utiliza WebRTC.
servers:
  - url: /
    description: Mesma origem do servidor / painel
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
  - connectionApiKeyAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT retornado por /api/auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Chave mestre configurada em KALLIA_API_KEY
    connectionApiKeyAuth:
      type: apiKey
      in: header
      name: X-Connection-API-Key
      description: Chave individual da conexão (kc_...)
...
  schemas:
    User:
      type: object
      properties:
        id: { type: string, example: "524c251a92e3a92d" }
        email: { type: string, example: "usuario@dominio.com" }
        role: { type: string, enum: [appadmin, admin, normal], example: "admin" }
        projectId: { type: string, example: "9a3fc0da79e330d9" }
...
paths:
...
  /api/auth/register:
    post:
      tags: [Autenticação]
      summary: Registra um novo usuário no projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: "novo@dominio.com" }
                password: { type: string, example: "123456" }
      responses:
        "200":
          description: Usuário cadastrado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "created" }
  /api/auth/login:
    post:
      tags: [Autenticação]
      summary: Realiza login do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: "admin@dominio.com" }
                password: { type: string, example: "123456" }
      responses:
        "200":
          description: Login efetuado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
...